requirements.txt
Django==2.1.2
django-crispy-forms==1.7.2
Pillow==5.3.0
pytz==2018.5
gunicorn==20.0.4
django-heroku==0.3.1
django-embed-video==1.3.2
django-paypal==1.0.0